Renamed agent main. #296
